添加音乐MV图片字段

+++ src/main/java/service/user/administrators/MailInformationService.java
@@ -40,58 +40,75 @@ public class MailInformationService {
 
     /**
      * 显示邮件信息
-     *
-     * @param condition 条件可以有多个 1、按id查询 2、按接收方id查询 3、按发送方id查询  4、按指定状态
+     *  类型-id-接收方-发送方—接收者
+     * @param condition 条件可以有多个 1、按id查询 2、按接收方id查询 3、按发送方id查询  4、按指定邮件等级
      * @param pageNum   表示当前第几页
-     * @param session   用于判断等级
+     * @param session   用于判断用户等级权限
      */
-    public String  showMail(String[] condition, Integer pageNum, HttpSession session,Model model) {
+    public String showMail(String[] condition, Integer pageNum, HttpSession session, Model model) {
         User user = specialFunctions.getUser(session);
         Mail mail = new Mail();
         if (condition != null) {
             if ((condition[0] != null) && !"".equals(condition[0])) {
-                mail.setId(Integer.parseInt(condition[0]));
-            }
-            if ((condition[1] != null) && !"".equals(condition[1])) {
-                mail.setRecipientId(Integer.parseInt(condition[1]));
+                // 1表示是id 2表示是接收方 3表示发送方
+                switch (condition[0]) {
+                    case "1":
+                        if ((condition[1] != null) && !"".equals(condition[1])) {
+                            mail.setId(Integer.parseInt(condition[1]));
+                        }
+                        break;
+                    case "2":
+                        condition[2]=condition[1];
+                        condition[1]=null;
+                        break;
+                    case "3":
+                        condition[3]=condition[1];
+                        condition[1]=null;
+                        break;
+                    default:
+                }
             }
             if ((condition[2] != null) && !"".equals(condition[2])) {
-                mail.setSenderId(Integer.parseInt(condition[2]));
+                mail.setRecipientId(Integer.parseInt(condition[2]));
             }
             if ((condition[3] != null) && !"".equals(condition[3])) {
-                mail.setState(Integer.parseInt(condition[3]));
+                mail.setSenderId(Integer.parseInt(condition[3]));
+            }
+            if ((condition[4] != null) && !"".equals(condition[4])) {
+                mail.setReply(Integer.parseInt(condition[4]));
             }
             model.addAttribute("conditionZero", condition[0]);
             model.addAttribute("conditionOne", condition[1]);
             model.addAttribute("conditionTwo", condition[2]);
             model.addAttribute("conditionThree", condition[3]);
+            model.addAttribute("conditionFour", condition[4]);
         } else {
             model.addAttribute("conditionZero", null);
             model.addAttribute("conditionOne", null);
             model.addAttribute("conditionTwo", null);
             model.addAttribute("conditionThree", null);
+            model.addAttribute("conditionFour", null);
         }
-//        // 控制显示的邮箱等级
-        if (user.getLevel() == 3) {
-            mail.setReply(1);
-        } else if (user.getLevel() >= 4) {
-            mail.setReply(2);
-        }
+//        // 控制显示的邮箱等级,邮件查看权限认证
+//        if (user.getLevel() == 3) {
+//            mail.setReply(1);
+//        } else if (user.getLevel() >= 4) {
+//            mail.setReply(2);
+//        }
         //在查询之前传入当前页，然后多少记录
-        PageHelper.startPage(pageNum, 9);
+        PageHelper.startPage(pageNum, 7);
         // 根据条件查找用户信息
         List<Mail> list = mailMapper.selectListMail(mail);
-        logger.debug("邮箱的信息"+list);
+        logger.debug("邮箱的信息" + list);
         // 传入页面信息
-        model.addAttribute("pageInfo",new PageInfo<>(list));
-        return "back_system/Email";
+        model.addAttribute("pageInfo", new PageInfo<>(list));
+        return "system/backgroundSystem";
     }
+
     /**
      * 显示指定id的邮箱信息
      */
-    @RequestMapping(value = "/showIdMail")
-    @ResponseBody
-    public Mail showIdMail(Integer id){
+    public Mail showIdMail(Integer id) {
         Mail mail = new Mail();
         mail.setId(id);
         List<Mail> list = mailMapper.selectListMail(mail);
@@ -149,13 +166,24 @@ public class MailInformationService {
     /**
      * 删除指定邮箱
      */
-    public String  deleteMail(Integer id,Model model,HttpSession session) throws DataBaseException {
-        // 删除指定id的评论
-        if (mailMapper.deleteMail(id) < 1) {
-            // 如果失败是数据库错误
-            logger.error("删除指定邮箱时，数据库出错");
-            throw new DataBaseException("删除指定邮箱时，数据库出错");
+    public State deleteMail(Integer id) throws DataBaseException {
+        State state = new State();
+        // 先判断id是否存在
+        // 根据条件查找邮件信息
+        Mail mail = new Mail();
+        mail.setId(id);
+        List<Mail> list = mailMapper.selectListMail(mail);
+        if (list.size() != 0) {
+            // 删除指定id的评论
+            if (mailMapper.deleteMail(id) < 1) {
+                // 如果失败是数据库错误
+                logger.error("删除指定邮箱时，数据库出错");
+                throw new DataBaseException("删除指定邮箱时，数据库出错");
+            }
+            state.setState(1);
+        } else {
+            state.setInformation("id不存在");
         }
-        return showMail(null,1,session,model);
+        return state;
     }
 }
